requirements.txt
streamlit
numpy
matplotlib
google-cloud-firestore
arrow
